+++ demo/server/index.js
@@ -1,6 +1,7 @@
 import { LitElement, html, css } from 'lit-element'
 import './db'
 import './api'
+import './realtime'
 
 customElements.define('demo-server', class extends LitElement{
 
@@ -22,6 +23,7 @@ customElements.define('demo-server', class extends LitElement{
         <b-tabs-router path="server/" key="server" layout="left" >
             demo-server-db
             demo-server-api
+            demo-server-realtime
         </b-tabs-router>
     `}