+++ realtime/server/views.js
@@ -2,9 +2,9 @@ const View = require('./view')
 
 class Views extends Map {
 	
-	connect(socket){
+	connect(io, socket){
 
-        this.io = socket.io
+        this.io = io
 
 		socket.on('view:open', (viewName, data)=>{
 			this.get(viewName).add(socket, data)
@@ -48,7 +48,7 @@ class Views extends Map {
 		let view = super.get(viewName)
 
 		if( !view ){
-			view = new View(io, viewName)
+			view = new View(this.io, viewName)
 			this.set(viewName, view)
 		}